feat: Update movement columns to numeric type in heroes table and adjust related logic

+++ supabase/schema.sql
@@ -63,8 +63,8 @@ create table public.heroes (
   defense integer not null default 1,
   spell_power integer not null default 1,
   knowledge integer not null default 1,
-  movement integer not null default 10,
-  max_movement integer not null default 10,
+  movement numeric not null default 10,
+  max_movement numeric not null default 10,
   x integer not null,
   y integer not null,
   is_moving boolean not null default false
